Now epty files search returns recent files
